% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_generator.R
\name{df_generator}
\alias{df_generator}
\title{df_generator.R}
\usage{
df_generator(
  rows,
  date_col = NULL,
  date_range = c(as.Date("2020-01-01"), as.Date("2025-01-01")),
  int_col = NULL,
  int_range = c(1, 100),
  float_col = NULL,
  float_min = 1,
  float_max = 100,
  boolean_col = NULL,
  letters_col = NULL,
  distribution_col = NULL,
  dist_type = "normal",
  dist_params = list(mean = 50, sd = 10),
  user_col = NULL,
  user_data = NULL,
  unique_user_col = NULL,
  unique_user = NULL,
  missing = NULL,
  n_missing = 0,
  id_col = NULL,
  unique_id = NULL,
  id_digits = 4,
  id_prefix = ""
)
}
\arguments{
\item{rows}{The number of rows in the data frame}

\item{cols}{The number of columns in the data frame}
}
\value{
specifically generated data frame
}
\description{
df_generator.R
}
