% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject_NA.R
\name{inject_NA}
\alias{inject_NA}
\title{inject_NA.R}
\usage{
inject_NA(df, columns, missing_percentage = 10)
}
\arguments{
\item{df}{The dataframe}

\item{columns}{Which columns to add missing values}
}
\value{
dataframe with injected NAs
}
\description{
inject_NA.R
}
