% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_df.R
\name{random_df}
\alias{random_df}
\title{random_df.R}
\usage{
random_df(rows, cols)
}
\arguments{
\item{rows}{The number of rows in the data frame}

\item{cols}{The number of columns in the data frame}
}
\value{
randomly generated data frame
}
\description{
random_df.R
}
