% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_generator.R
\name{df_generator}
\alias{df_generator}
\title{df_generator.R}
\usage{
df_generator(
  rows,
  date_col = NULL,
  date_range = c(as.Date("2020-01-01"), as.Date("2025-01-01")),
  int_col = NULL,
  int_range = c(1, 100),
  float_col = NULL,
  float_min = 1,
  float_max = 100,
  boolean_col = NULL,
  letters_col = NULL,
  distribution_col = NULL,
  dist_type = "normal",
  dist_params = list(mean = 50, sd = 10),
  user_col = NULL,
  user_data = NULL,
  unique_user_col = NULL,
  unique_user = NULL,
  missing = NULL,
  n_missing = 0,
  id_col = NULL,
  unique_id = NULL,
  id_digits = 4,
  id_prefix = ""
)
}
\arguments{
\item{rows}{The number of rows in the data frame}
\item{date_col}{date column name}
\item{date_range}{range for date_col}
\item{int_col}{integer column name}
\item{int_range}{range for int_col}
\item{float_col}{name for float column}
\item{float_min}{min for float column}
\item{float_max}{max for float column}
\item{boolean_col}{name for boolean column}
\item{letters_col}{name for letters col}
\item{distribution_col}{name for distribution column}
\item{dist_type}{distribution type: options are normal, binomial, or uniform}
\item{dist_params}{parameters for distribution column}
\item{user_col}{name for user_col}
\item{user_data}{data to be sampled (with replacement)}
\item{unique_user_col}{name for unique_user_col}
\item{unique_user}{data to be sampled (without replacement)}
\item{missing}{column to add missing data}
\item{n_missing}{number of missing datapoints to insert}
\item{id_col}{name for id_col}
\item{unique_id}{name for unique_id column}
\item{id_digits}{number of digits in the unique id}
\item{id_prefix}{prefix for unique id}
}
\value{
generated data frame
}
\description{
df_generator.R
}
